requirements.txt
opencv-python==4.11.0.86
mediapipe==0.10.21
scikit-learn==1.6.1
numpy==1.25.2
pyttsx3==2.98
Pillow==11.2.1

